+++ spb/graphics/__init__.py
@@ -18,7 +18,7 @@ from spb.graphics.complex_analysis import (
 )
 from spb.graphics.control import (
     control_axis, pole_zero, step_response, impulse_response, ramp_response,
-    bode_magnitude, bode_phase, nyquist, nichols
+    bode_magnitude, bode_phase, nyquist, nichols, root_locus, sgrid
 )
 
 __all__ = [
@@ -33,5 +33,5 @@ __all__ = [
     "contour_real_imag", "contour_abs_arg", "plane",
     "control_axis", "pole_zero", "step_response", "impulse_response",
     "ramp_response", "bode_magnitude", "bode_phase", "nyquist", "nichols"
-    "hline", "vline", "arrow_2d", "arrow_3d"
+    "hline", "vline", "arrow_2d", "arrow_3d", "root_locus", "sgrid"
 ]