+++ tests/backends/make_tests.py
@@ -4,7 +4,7 @@ from spb import (
     plot3d_parametric_line, plot3d, plot3d_list,
     plot3d_implicit, plot3d_parametric_surface,
     plot_vector, plot_complex, plot_real_imag, plot_riemann_sphere,
-    graphics, arrow_2d, arrow_3d
+    graphics, arrow_2d, arrow_3d, plot_root_locus
 )
 from spb.series import (
     SurfaceOver2DRangeSeries, ParametricSurfaceSeries, LineOver1DRangeSeries
@@ -1197,3 +1197,16 @@ def make_test_arrow_3d(B, lbl, rkw, sil):
             (1, 2, 3), (4, 5, 6), label=lbl, rendering_kw=rkw, show_in_legend=sil),
         show=False, backend=B, legend=True
     )
+
+
+def make_test_root_locus_1(B):
+    s = symbols("s")
+    G = (s**2 + 1) / (s**3 + 2*s**2 + 3*s + 4)
+    return plot_root_locus(G, backend=B, show=False)
+
+
+def make_test_root_locus_2(B):
+    s = symbols("s")
+    G1 = (s**2 + 1) / (s**3 + 2*s**2 + 3*s + 4)
+    G2 = (s**2 - 4) / (s**3 + 2*s - 3)
+    return plot_root_locus((G1, "a"), (G2, "b"), backend=B, show=False)